Don't wrap text at word boundaries, but rather anywhere.
Actually, this means, that text doesn't disappear any longer: We
don't provide space for wrapping. Previously, we used default options
which was some kind of word wrapping (Though Docs says, this has to
be enabled ?!?).
Now, we allow to "break anywhere", which means: Draw until the end of
the line, then wrap and stop drawing (there's no more space left)